refactor: migrate to domain-driven api layout
