+++ api/rest/__init__.py
@@ -1,0 +1,3 @@
+from . import auth, intentions, patches, sessions
+
+__all__ = ["auth", "sessions", "intentions", "patches"]
