+++ domains/auth/__init__.py
@@ -1,0 +1,20 @@
+from .models import Token, User, UserCreate, UserLogin
+from .service import (
+    AuthService,
+    check_session_access,
+    get_current_user,
+    grant_session_access,
+    init_auth_db,
+)
+
+__all__ = [
+    "AuthService",
+    "User",
+    "UserLogin",
+    "UserCreate",
+    "Token",
+    "get_current_user",
+    "grant_session_access",
+    "check_session_access",
+    "init_auth_db",
+]
