+++ domains/auth/exceptions.py
@@ -1,0 +1,4 @@
+class AuthenticationError(Exception):
+    """Raised when authentication fails."""
+
+    pass
