+++ domains/state/context_slicer.py
