+++ shared/database/__init__.py
@@ -1,0 +1,3 @@
+from .connection import get_db, init_db
+
+__all__ = ["get_db", "init_db"]
