+++ shared/database/migrations.py
@@ -1,0 +1,3 @@
+"""Database migration utilities (placeholder)."""
+
+# In a real project, use Alembic or similar tool.
