+++ shared/database/repositories/__init__.py
@@ -1,0 +1,11 @@
+from .session_repo import create_session, get_session
+from .state_repo import create_initial_state, get_state
+from .user_repo import get_user_by_email
+
+__all__ = [
+    "create_session",
+    "get_session",
+    "create_initial_state",
+    "get_state",
+    "get_user_by_email",
+]
